requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.5.0
async-timeout==4.0.3
asynctest==0.13.0
attrs==24.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cycler==0.11.0
dnspython==2.3.0
fonttools==4.38.0
frozenlist==1.3.3
idna==3.10
importlib-metadata==6.7.0
Jinja2==3.1.6
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.5.3
multidict==6.0.5
numpy==1.21.6
packaging==24.0
pandas==1.1.5
Pillow==9.5.0
pydantic==2.5.3
pydantic-settings==2.0.3
pydantic_core==2.14.6
pymongo==4.7.3
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==0.21.1
pytz==2025.1
PyYAML==6.0.1
six==1.17.0
structlog==23.1.0
tenacity==8.2.3
typing_extensions==4.7.1
XlsxWriter==3.2.2
yarl==1.9.4
zipp==3.15.0
